third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/cross-origin-isolated-permission.https-expected.txt

This is a testharness.js-based test.
PASS origin = https://web-platform.test:8444, value = undefined
PASS origin = https://web-platform.test:8444, value = *
PASS origin = https://web-platform.test:8444, value = 'self'
FAIL origin = https://web-platform.test:8444, value = 'none' assert_equals: expected false but got true
PASS origin = https://www1.web-platform.test:8444, value = undefined
FAIL origin = https://www1.web-platform.test:8444, value = * assert_equals: expected true but got false
PASS origin = https://www1.web-platform.test:8444, value = 'self'
PASS origin = https://www1.web-platform.test:8444, value = 'none'
Harness: the test ran to completion.
